It works! (if you only display the upper left of your world :-)

+++ phone/TileGrid.h
@@ -27,7 +27,7 @@
 	
 	GLfloat *tileCoordinates;
 	GLfloat *tileTextureCoordinates;
-	GLfloat *tileIndexes;
+	GLushort *tileIndexes;
 	GLushort numCoordinates;
 	GLushort numIndexes;
 }
